The ConveyorBelt.cs "powers" on when the targetGear powers on. Abstract the Power system into a PoweredObject.cs so we can easily reuse it.
