--- Assets/ConveyorBelt.cs
@@ -1,22 +1,0 @@
-﻿using UnityEngine;
-
-public class ConveyorBelt : MonoBehaviour
-{
-    /// <summary>
-    /// The object that controls movement.
-    /// </summary>
-    [SerializeField]
-    private GameObject triggerObject = null;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
