+++ Assets/Scripts/Logic/PoweredObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3516036d59e9fb0469e1e2c1d723f08b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
